// apps/web/src/lib/api/generated/models/programOutcome.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */
import type { Program } from "./program";

export interface ProgramOutcome {
  id: string;
  institution_id: string;
  program_id: string;
  type: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  syllabus_scheme: string | null;
  academic_year: number;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
  program?: Program;
}